requirements.txt
pandas==1.5.2
numpy==1.23.5
matplotlib==3.6.2
seaborn==0.11.2
scikit-learn==1.2.0
torch==1.13.0
torchvision==0.14.0
opencv-python==4.6.0.66
Pillow==9.3.0